Fix and improve message ID handling
The extractor and the plugin used a slightly different method when it
came to extracting the message ID from a simple text. One would include
leading and trailing whitespace, while the other did not. Now both leave
leading and trailing whitespace out of the message ID.

Leading and trailing whitespace was not stripped at all from the message
ID for complex messages, but that could lead to a poor translator
experience when having to handle the PO(T)-files, so complex messages
now also leave the leading and trailing whitespace out of the message
ID.